+++ src/CSharpLanguageServer/DocumentationUtil.fs
@@ -99,7 +99,11 @@ module DocumentationUtil =
             { comment with Returns = comment.Returns |> List.append [ n ] }
 
         | "exception" ->
-            let name = n.Attribute(XName.Get("cref")).Value |> parseCref
+            let name = n.Attribute(XName.Get("cref"))
+                       |> Option.ofObj
+                       |> Option.map (fun a -> parseCref a.Value)
+                       |> Option.defaultValue "(unspecified)"
+
             { comment with
                 Exceptions = comment.Exceptions |> List.append [ (name, n) ] }