+++ CHANGELOG.md
@@ -3,6 +3,11 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [Unreleased]
+* Fix completion item kind for an extension method
+  - Reported by @sharpchen in https://github.com/razzmatazz/csharp-language-server/issues/237
+  - Fixed in https://github.com/razzmatazz/csharp-language-server/pull/238
+
 ## [0.18.0] - 2025-06-23 / Pabradė
 * Upgrade Roslyn to 4.14.0
 * Use diagnostic url from Roslyn