+++ CHANGELOG.md
@@ -7,6 +7,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 * Fix completion item kind for an extension method
   - Reported by @sharpchen in https://github.com/razzmatazz/csharp-language-server/issues/237
   - Fixed in https://github.com/razzmatazz/csharp-language-server/pull/238
+* Documentation formatting fixes
+  - Reported by @prashanthbabu07 in https://github.com/razzmatazz/csharp-language-server/issues/247
+  - Fixed in https://github.com/razzmatazz/csharp-language-server/pull/248
 
 ## [0.18.0] - 2025-06-23 / Pabradė
 * Upgrade Roslyn to 4.14.0