+++ CHANGELOG.md
@@ -3,7 +3,7 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
-## [Unreleased]
+## [0.19.0] - 2025-08-20 / Kapčiamiestis
 * Select common target framework when multiple projects are found
   - By @AdeAttwood in https://github.com/razzmatazz/csharp-language-server/pull/253
 * Fix how completion item details are resolved